Add component modules and update tests
