+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": [
     "env",
+    "stage-0",
     "react"
   ]
 }
