+++ package.json
@@ -7,6 +7,7 @@
     "prepublish": "babel src -d dist",
     "start": "start-storybook -p 8000 -c stories",
     "demo": "x0 dev docs/Demo.js -op 9000",
+    "xdemo": "x0 build docs/Demo.js --static -d demo",
     "storybook": "build-storybook -c stories -o docs/stories",
     "cover": "nyc report --reporter=html --reporter=lcov && codecov",
     "test": "nyc ava"
@@ -40,6 +41,7 @@
     "babel-core": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
     "babel-register": "^6.26.0",
     "codecov": "^3.0.0",
     "nyc": "^11.4.1",
