+++ src/Circle.js
@@ -1,4 +1,5 @@
 import sys from 'system-components'
+import Badge from './Badge'
 import { sizeWidth, sizeHeight } from './utils'
 
 export const Circle = sys({
